feat: completed the death animation stuff and switched my map's assets over to the Terrain Layer
The enemies now sink into the ground and shrink as they are doing it.  Also, that stats for how that works are exposed and changeable in the inspector.
Next I will add that same lgoci to the spawn so enemies come up into the field instead of popping into existence.