feat: Added all enemy audio code except for spawn and proximity related stuff
I don't know how to add the spawn audio because I don't know when the enemies spawn in the code.  Also, I don't know how to add the proximity sound alert as I realized it looks harder than I initially gave it credit for.
I added the Attack Audio code to the EnemyAttackStats, but not AbilityStats since that is outside my jurisdiction.