Qa branch (#60)
* feat: added to the enemy melee attack script saying that if the enemy is dead, just turn off the melee collider.

* feat: fixed the Overworld Portal's Ring to be interactable without having to look at the ring.  Changed the Snake and Spike to be bigger and have colliders that work better.

* feat: reduced the size of the melee attack colliders on the enemies.

* feat: started implementing the new avoidance system.  Added this to the Wolf, Spike, Snake, Hermit King Boss, and most importantly, the Player.  I also created the layer needed for the new avoidance system to work.

Everything that needs the avoidance system to apply to needs a child object on the "Avoidance" layer and the "No Friction" physic on the object's collider.  Also, rename that object "Avoidance Collider".

* feat: had to re-place a lot of the materials for the Boss Room as they had gotten removed somehow.

* feat: added avoidance colliders to the Boss Room as well as the Pillar prefabs

This does not work currently.  The Hermit King does not seem to care about Avoidance.  Needs further research.

* feat: added the new Avoidance to Bloom, Blossom, Fire Dragon, Inferno Dragon Final Boss, and Poison Bomb.

I tested Poison Bomb, and it's avoidance failed.  It might be because that enemy is not complete, but I don't know.  I haven't tested the others, but I will after this commit.

* feat: Added a new Scene to replace my Sandbox which has something wrong with it.  Also, made slight changes to the wolf and Pillar prefabs.

The Wolf's avoidance's mesh renderer was turned on.  I increased the size of the avoidance of the Pillars to try to get the Hermit King Boss to cooperate.  It didn't work.

* feat: updated the NavMesh of the Boss Room.

This will keep the Bosses from "entering" the walls and pillars.

* feat: reshrank the avoidance of the Pillars so it isnt visible

* feat: updated the stoppingdistance of the snake to stay away from the player, updated the pillar to make the base unwalkable, readjusted the Boss Room navmesh, removed the avoidance from the player and Boss Room Walls, and added the boundary walls and floor to the Boss Room.

* feat: updated the Boss Room navmesh to make the boss properly avoid pillars while still being able to melee attack the player